SuperMean/backend/requirements.txt
# Directory: backend/
# File: requirements.txt
# Description: Python package requirements for the backend project. (Updated)

# Core Web Framework & Server
fastapi>=0.95.0,<0.111.0
uvicorn[standard]>=0.20.0,<0.25.0

# Data Validation & Settings
pydantic>=1.10.0,<2.0.0
pydantic-settings>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
PyYAML>=6.0

# Logging
loguru>=0.7.0 # Or remove if only using standard logging

# HTTP Requests (for external APIs, skills, DeepSeek)
requests>=2.28.0

# --- LLM & AI ---
# Google Gemini
google-generativeai>=0.3.0

# OpenAI library (Used for AIMLAPI Router Connector)
openai>=1.0.0 # Use version 1.x+ for new API structure

# SerpApi (for Web Search Skill)
google-search-results>=2.4.0

# Placeholder for DeepSeek SDK (if/when available via pip)
# deepseek-sdk==x.y.z

# Placeholder for other potential LLM SDKs (e.g., Anthropic)
# anthropic==x.y.z

# --- Memory ---
# Placeholder for Vector Database client (choose one or more)
# faiss-cpu
# faiss-gpu
# pinecone-client
# qdrant-client
# chromadb

# --- Optional but potentially useful ---
# Job Queues
# celery==x.y.z
# redis==x.y.z

# Database ORM
# sqlalchemy==x.y.z
# databases

# Testing (install separately)
# pytest
# pytest-asyncio
# httpx